+++ src/client/renderer/r_atlas.h
@@ -31,10 +31,10 @@ void R_CompileAtlas(r_atlas_t *atlas);
 #ifdef __R_LOCAL_H__
 
 typedef struct {
-	uint32_t width;
-	uint32_t height;
-	uint32_t x;
-	uint32_t y;
+	uint16_t width;
+	uint16_t height;
+	uint16_t x;
+	uint16_t y;
 
 	uint32_t right;
 	uint32_t down; // nodes to the right/below this one
@@ -49,17 +49,17 @@ typedef struct {
 	// required for writing
 	uint32_t num_alloc_nodes;
 	_Bool keep_square;
-	uint32_t max_width, max_height;
+	uint16_t max_width, max_height;
 } r_packer_t;
 
-void R_AtlasPacker_InitPacker(r_packer_t *packer, const uint32_t max_width, const uint32_t max_height,
-                              const uint32_t root_width, const uint32_t root_height, const uint32_t initial_size);
+void R_AtlasPacker_InitPacker(r_packer_t *packer, const uint16_t max_width, const uint16_t max_height,
+                              const uint16_t root_width, const uint16_t root_height, const uint32_t initial_size);
 void R_AtlasPacker_FreePacker(r_packer_t *packer);
-r_packer_node_t *R_AtlasPacker_GrowNode(r_packer_t *packer, const uint32_t width, const uint32_t height);
-r_packer_node_t *R_AtlasPacker_SplitNode(r_packer_t *packer, r_packer_node_t *node, const uint32_t width,
-        const uint32_t height);
-r_packer_node_t *R_AtlasPacker_FindNode(r_packer_t *packer, const uint32_t root, const uint32_t width,
-                                        const uint32_t height);
+r_packer_node_t *R_AtlasPacker_GrowNode(r_packer_t *packer, const uint16_t width, const uint16_t height);
+r_packer_node_t *R_AtlasPacker_SplitNode(r_packer_t *packer, r_packer_node_t *node, const uint16_t width,
+        const uint16_t height);
+r_packer_node_t *R_AtlasPacker_FindNode(r_packer_t *packer, const uint32_t root, const uint16_t width,
+                                        const uint16_t height);
 void R_AtlasPacker_Serialize(const r_packer_t *packer, file_t *file);
 _Bool R_AtlasPacker_Unserialize(file_t *file, r_packer_t *packer);
